Adjusted typing around connection strings in Pongo
