+++ src/packages/dumbo/src/core/connections/pool.ts
@@ -76,10 +76,13 @@ export const createConnectionPool = <
   return result as ConnectionPoolType;
 };
 
-export const createDeferredConnectionPool = <Connector extends ConnectorType>(
+export const createDeferredConnectionPool = <
+  Connector extends ConnectorType,
+  ConnectionType extends Connection<Connector> = Connection<Connector>,
+>(
   connector: Connector,
   importPool: () => Promise<ConnectionPool<Connection<Connector>>>,
-): ConnectionPool<Connection<Connector>> => {
+): ConnectionPool<ConnectionType> => {
   let poolPromise: Promise<ConnectionPool<Connection<Connector>>> | null = null;
 
   const getPool = async (): Promise<ConnectionPool<Connection<Connector>>> => {
