+++ src/packages/pongo/src/storage/all/index.ts
@@ -1,18 +1,24 @@
-import type { SupportedDatabaseConnectionString } from '@event-driven-io/dumbo';
+import type {
+  ConnectorType,
+  DatabaseConnectionString,
+  InferConnectorDatabaseType,
+} from '@event-driven-io/dumbo';
 import { NodePostgresConnectorType } from '@event-driven-io/dumbo/pg';
-import type { PongoClientOptions } from '../../core';
-import type { PostgresDbClientOptions } from '../postgresql';
+import type { PongoClientOptions, PongoDbClientOptions } from '../../core';
 
 export const clientToDbOptions = <
-  ConnectionString extends SupportedDatabaseConnectionString,
+  ConnectionString extends DatabaseConnectionString<
+    InferConnectorDatabaseType<Connector>
+  >,
+  Connector extends ConnectorType = ConnectorType,
   DbClientOptions extends
-    PostgresDbClientOptions<ConnectionString> = PostgresDbClientOptions<ConnectionString>,
+    PongoDbClientOptions<ConnectionString> = PongoDbClientOptions<ConnectionString>,
 >(options: {
   connectionString: ConnectionString;
   dbName?: string;
   clientOptions: PongoClientOptions<ConnectionString>;
 }): DbClientOptions => {
-  const postgreSQLOptions: PostgresDbClientOptions<ConnectionString> = {
+  const postgreSQLOptions: PongoDbClientOptions<ConnectionString> = {
     connector: NodePostgresConnectorType,
     dbName: options.dbName,
     ...options.clientOptions,
